+++ packages/vendure-plugin-invoices/src/strategies/storage/local-file-strategy.ts
@@ -1,7 +1,7 @@
 import { Response } from 'express';
 import { createReadStream, promises as fs, ReadStream } from 'fs';
-import { InvoiceEntity } from '../entities/invoice.entity';
-import { exists, zipFiles, ZippableFile } from '../util/file.util';
+import { InvoiceEntity } from '../../entities/invoice.entity';
+import { exists, zipFiles, ZippableFile } from '../../util/file.util';
 import { LocalStorageStrategy } from './storage-strategy';
 
 /**