+++ packages/vendure-plugin-invoices/src/strategies/storage/s3-storage.strategy.ts
@@ -1,12 +1,12 @@
 import { Logger } from '@vendure/core';
 import { readFile, writeFile } from 'fs/promises';
-import { InvoiceEntity } from '../entities/invoice.entity';
+import { InvoiceEntity } from '../../entities/invoice.entity';
 import {
   safeRemove,
   ZippableFile,
   createTempFile,
   zipFiles,
-} from '../util/file.util';
+} from '../../util/file.util';
 import { createReadStream, ReadStream } from 'fs';
 import { RemoteStorageStrategy } from './storage-strategy';